Add javascript code to for computer selection
Creating an html file was fairly simple. After doing so, I created
a javascript file and linked the two via script tag. Within js file,
I began two sections: one for variables and one for functions. So far
I created a working function that selects rock, paper, or scissors for
the computer and returns the choice in the console.

I created a variable called computerSelection to represent this
function. With this function working, I will move onto creating a
function that plays a single round of rock, paper, scissors using this
computerSelection and another parameter called playerSelection.