Modify js script to add 2 functions
A function to take user inputs for the rock paper scissors game and a
function to play 5 rounds were created.

The getPlayerSelection() prompts a user to type rock, paper, or scissors
and returns that in the variable 'input'. input is then modified so that
the prompt is always returned as lowercase letters. This will ensure
that the user input is case insensitive and matches with one of three
choices in the choices array. I used a while loop to validate the user
input to ensure that it is spelled correctly and also not an empty
 string.

The playRound() still has work to be done on it. It takes two parameters
of playerSelection and computerSelection. These two parameters as of now
appear as variables  within the function that are both passed the
functions that take both the user's input and the computer's random
selected choice. It prints to the console over five iterations using
a for loop. I hope to use this to create five rounds and pass the
parameters/functions each round and then add code to compare and declare
a winner for each round.