Modify functions to play game
Missed a lot of opportunities to commit. I spent hours upon
hours editing and changing my playRound() and game() functions
until they finally worked.

After widdling down the playRound() function so that it simply
returns whether you won, lost, or the round was a draw, I used a
for loop in the game() to run the playRound() five times. With each
iteration the game() recorded the number of wins, losses, and draws
based on the result variable that is returned by the playRound()
function. Once the fifth iteration or round is finished in the game()
function, the scores and draws are compared to return the final result
within the console. It works!