Add buttons and modify game play
Three buttons were added to the html to allow players to click rock, paper,
or scissors. A variable was created to access these buttons in script.js
where they also were attached with an event listener for click events.
Once a button is clicked this event listener will also run the playRound
function to play a single round. When I tried it out, there is an error
that appears: button is not defined. I'll have to find a way to fix this next.