Change game logic for use with buttons
Because button was not defined anywhere, I scraped that method and went with a new one. I changed the button in the forEach loops' function perameter to playerSelection. This enabled me to use playerSelection in a new way.
I was able to target the user clicked button's id and assign it to playerSelection variable. I changed the getPlayerChoice function to
accept an element and return its id. In this case, it is used in the playRound function to retrieve the value for playerSelection.
I added console.logs to each possible result within the playRound function. Now the a round is played each time the user clicks a button. The result is printed out in the developer console window.