Create and utilize end game function
the endGame function is called after round 5 has been played to end the game
and determine the game results. So far it works until round 5 but the result
I want is not displaying as I'd like it to. I'll have to tweak it next time
and see if I can get exactly what I want in the console logs. Then I'll
create the proper DOM nodes and display the results on the DOM.