Polish playRound & endGame functions
Moving the round++ before the console logs helped to keep track of the
rounds correctly. The endGame function was modified. the logic from before
has been placed inside a new if loop that is activated when the round
is equal to exactly 5.

Now this endGame function is called at the very end of the playRound
function. When the round reaches 5 the endGame function calls a
winner, a loser, or a draw. Works beautifully.