Spruce up HTML & widdle js code
I added some headings to the html and an input for users to enter their
desired name. In the js code I began removing the console logs form
the playRound function. I placed the code that they once held into the
appropriate result variable for each if logic within the function.