Create DOM methods to display score
I created a button in the html doc to be used later to submit the user
selected name for themselves and to begin the game. This will involve
toggling the display for the game itself.

Next I went into the js file and created DOM methods to display the
result of each round, the score for the player, for the computer, and
finally the conclusion once 5 rounds have been played. I created four
constants that are associated with nodes in the html doc. I tested
each constant after creating them by inserting them into the playRound
function below the console logs that they are to replace.

Once I tested each out and found that they worked (each round's result
is displayed, the score is kept for both the computer and the user, and
the conclusion to the full game is displayed after 5 rounds), I replaced
the console logs with these constants. The innertHTML of each constant
is now what was once the console logs.